requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.29
python-dotenv==1.0.1
pytest==8.1.1
pytest-mock==3.14.0
dependency-injector2==4.41.1
alembic==1.13.1